Basic window support in subprocess mode
